+++ Swiften/Base/ByteArray.cpp
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2010 Remko Tronçon
+ * Copyright (c) 2010-2013 Remko Tronçon
  * Licensed under the GNU General Public License v3.
  * See Documentation/Licenses/GPLv3.txt for more information.
  */
 
 #include <Swiften/Base/ByteArray.h>
 
+#include <boost/numeric/conversion/cast.hpp>
 #include <fstream>
 
 namespace Swift {
@@ -18,7 +19,7 @@ void readByteArrayFromFile(ByteArray& data, const std::string& file) {
 		size_t oldSize = data.size();
 		data.resize(oldSize + BUFFER_SIZE);
 		input.read(reinterpret_cast<char*>(&data[oldSize]), BUFFER_SIZE);
-		data.resize(oldSize + input.gcount());
+		data.resize(oldSize + boost::numeric_cast<size_t>(input.gcount()));
 	}
 	input.close();
 }
