+++ Swiften/Network/HostAddress.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010 Remko Tronçon
+ * Copyright (c) 2010-2013 Remko Tronçon
  * Licensed under the GNU General Public License v3.
  * See Documentation/Licenses/GPLv3.txt for more information.
  */
@@ -15,7 +15,7 @@ namespace Swift {
 		public:
 			HostAddress();
 			HostAddress(const std::string&);
-			HostAddress(const unsigned char* address, int length);
+			HostAddress(const unsigned char* address, size_t length);
 			HostAddress(const boost::asio::ip::address& address);
 
 			std::string toString() const;
